minimega: split client code into separate package.
Split the client code into it's own package so that it can be reused.
Going to reuse it for test framework.
